refactor(billing): remove temporary usage barrel
Replace the remaining billing usage barrel imports with direct module imports so billing helpers live behind their owning modules instead of a compatibility layer. Update the related PRD and todo docs to record that the transitional barrel has been deleted.
