+++ src/app/api/v1/plans/stream/helpers.ts
@@ -9,7 +9,7 @@ import type { GenerationResult } from '@/features/ai/types/orchestrator.types';
 import type { ParsedModule } from '@/features/ai/types/parser.types';
 import type { StreamingEvent } from '@/features/ai/types/streaming.types';
 import type { GenerationAttemptResult } from '@/features/plans/lifecycle/types';
-import { incrementUsage } from '@/features/billing/usage';
+import { incrementUsage } from '@/features/billing/usage-metrics';
 import { getCorrelationId } from '@/lib/api/context';
 import type { AttemptsDbClient } from '@/lib/db/queries/types/attempts.types';
 import { getDb } from '@/lib/db/runtime';
